pyproject.toml
[project]
name = "crypto-momentum"
version = "0.1.0"
description = "Crypto time-series momentum backtesting system"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "scipy>=1.11",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 85
